+++ test/AuthorizationMiddlewareFactoryTest.php
@@ -1,19 +1,20 @@
 <?php
+
 /**
- * @see       https://github.com/zendframework/zend-expressive-authorization for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio-authorization for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authorization/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authorization/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Expressive\Authorization;
+namespace MezzioTest\Authorization;
 
+use Mezzio\Authorization\AuthorizationInterface;
+use Mezzio\Authorization\AuthorizationMiddleware;
+use Mezzio\Authorization\AuthorizationMiddlewareFactory;
+use Mezzio\Authorization\Exception;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
-use Zend\Expressive\Authorization\AuthorizationInterface;
-use Zend\Expressive\Authorization\AuthorizationMiddleware;
-use Zend\Expressive\Authorization\AuthorizationMiddlewareFactory;
-use Zend\Expressive\Authorization\Exception;
 
 class AuthorizationMiddlewareFactoryTest extends TestCase
 {
@@ -35,6 +36,7 @@ class AuthorizationMiddlewareFactoryTest extends TestCase
     public function testFactoryWithoutAuthorization()
     {
         $this->container->has(AuthorizationInterface::class)->willReturn(false);
+        $this->container->has(\Zend\Expressive\Authorization\AuthorizationInterface::class)->willReturn(false);
 
         $this->expectException(Exception\InvalidConfigException::class);
         $middleware = ($this->factory)($this->container->reveal());